+++ changelog.md
@@ -15,6 +15,7 @@ Patch: Everything else → bumps patch version (1.0.0 → 1.0.1)
 - Area for notices below the "Come worship with us" section for short term updates
 - Notices can be stored in and loaded from database
 - Notices are filtered for the most current notice
+- Admin page for notices (admin/noticesadmin)
 
 ### Changed